+++ changelog.md
@@ -1,4 +1,5 @@
 ##### Version 0.16-109
+- Fixed recipes not loading on dedicated servers (BluSunrize)
 - Also includes the changes from 0.14-107
 
 ##### Version 0.15-108